// src/api/openapi/model/authResponse.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


export interface AuthResponse { 
    accessToken: string;
    refreshToken: string;
    user: User;
}